+++ plugin/src/main/kotlin/com/jetbrains/packagesearch/plugin/PackageSearchToolWindowFactory.kt
@@ -7,9 +7,9 @@ import com.intellij.openapi.wm.ToolWindow
 import com.intellij.openapi.wm.ToolWindowFactory
 import com.jetbrains.packagesearch.plugin.ui.LocalComponentManager
 import com.jetbrains.packagesearch.plugin.ui.PackageSearchToolwindow
-import com.jetbrains.packagesearch.plugin.ui.panels.packages.PackageSearchTabStyle
-import com.jetbrains.packagesearch.plugin.ui.panels.packages.PackageSearchTreeStyle
-import com.jetbrains.packagesearch.plugin.ui.panels.packages.PackageSearchGlobalColors
+import com.jetbrains.packagesearch.plugin.ui.bridge.PackageSearchGlobalColors
+import com.jetbrains.packagesearch.plugin.ui.bridge.PackageSearchTabStyle
+import com.jetbrains.packagesearch.plugin.ui.bridge.PackageSearchTreeStyle
 import com.jetbrains.packagesearch.plugin.utils.installActions
 import org.jetbrains.jewel.bridge.addComposeTab
 import org.jetbrains.jewel.foundation.LocalGlobalColors
